// shared/src/types/network/websocket/wsMeetingEvents.ts
type BasicMeetingEvent = {
	sentDate: string;
	meetingId: string;
};
// ...
export type MeetingAudioStreamChangedEvent = BasicMeetingEvent & {
	type: WsEventType.MEETING_AUDIO_STREAM_CHANGED;
	active: boolean;
	userId: string;
	moderatorId?: string;
};